LayoutTests/platform/ios/fast/canvas/pattern-too-large-to-create-expected.txt

CONSOLE MESSAGE: line 18: Canvas area exceeds the maximum limit (width * height > 16777216).
PASS: Saw exception.
